Add projects list and make Table component
Add projects list to constants file, and use that list in Table component. Use Table component for rendering projects list. Seperate Table component for better readability and code
