+++ src/Pages/Projects.js
@@ -1,0 +1,13 @@
+import React from "react";
+import { Table } from "../Components/Table";
+
+const Projects = () => {
+  return (
+    <div className="flex justify-center items-center mt-12">
+      <div className="w-[600px] min-w-[400px] border border-gray-100 shadow-lg">
+        <Table />
+      </div>
+    </div>
+  );
+};
+export default Projects;
