+++ src/constants/projects.js
@@ -1,0 +1,37 @@
+export const projects = [
+    {
+        projectName: "Tech 2024",
+        projectLink: "https://google.com",
+        transcriptionFile: "transcription.txt",
+        subtitleFile: "subtitle.vtt",
+        thumbailImg: "thumbnail.png"
+    },
+    {
+        projectName: "Tech 2024",
+        projectLink: "https://google.com",
+        transcriptionFile: "transcription.txt",
+        subtitleFile: "subtitle.vtt",
+        thumbailImg: "thumbnail.png"
+    },
+    {
+        projectName: "Tech 2024",
+        projectLink: "https://google.com",
+        transcriptionFile: "transcription.txt",
+        subtitleFile: "subtitle.vtt",
+        thumbailImg: "thumbnail.png"
+    },
+    {
+        projectName: "Tech 2024",
+        projectLink: "https://google.com",
+        transcriptionFile: "transcription.txt",
+        subtitleFile: "subtitle.vtt",
+        thumbailImg: "thumbnail.png"
+    },
+    {
+        projectName: "Tech 2024",
+        projectLink: "https://google.com",
+        transcriptionFile: "transcription.txt",
+        subtitleFile: "subtitle.vtt",
+        thumbailImg: "thumbnail.png"
+    }
+]
